Backend/foodclassifier-service/app/requirements.txt
albumentations==1.2.1
torch==1.12.0
flask==2.1.0
flask-restx==0.5.1
flask_jwt_extended
torchvision==0.13.0
werkzeug==2.1.2
opencv-python
pandas==1.4.3
numpy==1.23.1
click
Jinja2
itsdangerous
MarkupSafe
uWSGI
pillow
boto3
python-logstash